add profile image and login logout button in header
