+++ lib/liveview_chat_web/controllers/auth_controller.ex
@@ -1,5 +1,6 @@
-defmodule LiveviewChatWeb.AuthLiveView do
-  import Phoenix.LiveView
+defmodule LiveviewChatWeb.AuthController do
+  use LiveviewChatWeb, :controller
+  import Phoenix.LiveView, only: [assign_new: 3]
 
   def on_mount(:default, _params, %{"jwt" => jwt} = _session, socket) do
     socket =
@@ -21,5 +22,16 @@ defmodule LiveviewChatWeb.AuthLiveView do
   def on_mount(:default, _params, _session, socket) do
     socket = assign_new(socket, :loggedin, fn -> false end)
     {:cont, socket}
+  end
+
+  def login(conn, _params) do
+    redirect(conn, external: AuthPlug.get_auth_url(conn, "/"))
+  end
+
+  def logout(conn, _params) do
+    conn
+    |> AuthPlug.logout()
+    |> put_status(302)
+    |> redirect(to: "/")
   end
 end
