--- lib/liveview_chat_web/controllers/login_controller.ex
@@ -1,14 +1,0 @@
-defmodule LiveviewChatWeb.LoginController do
-  use LiveviewChatWeb, :controller
-
-  def login(conn, _params) do
-    redirect(conn, external: AuthPlug.get_auth_url(conn, "/"))
-  end
-
-  def logout(conn, _params) do
-    conn
-    |> AuthPlug.logout()
-    |> put_status(302)
-    |> redirect(to: "/")
-  end
-end
