+++ mix.exs
@@ -56,7 +56,7 @@ defmodule LiveviewChat.MixProject do
       {:plug_cowboy, "~> 2.5"},
       {:excoveralls, "~> 0.14.4", only: [:test, :dev]},
       {:ping, "~> 1.1.0"},
-      {:auth_plug, "~> 1.4.8"}
+      {:auth_plug, "~> 1.4.10"}
     ]
   end
